requirements.txt
Django
Pillow
django-tgbot
python-dotenv
mysqlclient
gunicorn
